[fx graph cache] Refactor FxGraphCachePickler, step 2
Summary: Move all the custom _reduce functions inside the FxGraphCachePickler class. This is mostly a cosmetic change. But in an upcoming diff, I'll add a member variable to the class to dictate how constant tensors are pickled, so it will be convenient to be able to query that setting via `self`. I made the analogous changes to AOTAutogradCachePickler for consistency.

Test Plan: unit tests

ghstack-source-id: 8df1377c527a4b08cecd18683ef001aab9d27575
Pull Request resolved: https://github.com/pytorch/pytorch/pull/138683