[PyTorch] Migrate bf16 gemv fast path kernel from intrinsics to vec::Vectorized
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139159

Very similar to #137912, but for bf16. (This is building toward enabling this fast path on non-ARM architectures, and in particular on x86 for machines without AVX512BF16).

Testing: checked for regression with llm_experiments' `benchmarks/benchmark_torch_mm.py llm` on M1 Mac and it appeared to be neutral. Supported this assessment by inspecting assembly for the bf16 dot kernel; observed minor instruction scheduling changes but nothing more.

ghstack-source-id: 251386473
@exported-using-ghexport

Differential Revision: [D65120325](https://our.internmc.facebook.com/intern/diff/D65120325/)